Move readable stream default controller to separate file
